+++ server/prisma/migrations/20260909163431_permiso_fijar_etapa/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "Action" ADD VALUE 'FIJAR_ETAPA';